fix i_lower bug

+++ src/objective/objective_function.cpp
@@ -281,7 +281,7 @@ void InnerQuantize(const score_t value, int_score_t* out_value_int, double scale
       *out_value_int = int_values[i_found];
     }
   } else {
-    size_t i_lower = i_found - 1;
+    int i_lower = static_cast<int>(i_found) - 1;
     CHECK_GE(i_lower, 0);
     while (scaled_value < int_values[i_lower]) {
       CHECK_GT(i_lower, 0);
